+++ scripts/redact_cli_py/tests/factories/fott_label_factory.py
@@ -28,3 +28,21 @@ class FottLabelFactory:
             fott_label_dict = json.load(fott_label_json)
             fott_label = from_dict(data_class=FottLabel, data=fott_label_dict)
             return fott_label
+
+    @staticmethod
+    def build_partial() -> FottLabel:
+        fott_label_path = "testdata/testdata-partial.jpg.labels.json"
+
+        with open(fott_label_path, encoding='utf-8-sig') as fott_label_json:
+            fott_label_dict = json.load(fott_label_json)
+            fott_label = from_dict(data_class=FottLabel, data=fott_label_dict)
+            return fott_label
+
+    @staticmethod
+    def build_redacted_partial() -> FottLabel:
+        fott_label_path = "testdata/testdata-partial.redacted.labels.json"
+
+        with open(fott_label_path, encoding='utf-8-sig') as fott_label_json:
+            fott_label_dict = json.load(fott_label_json)
+            fott_label = from_dict(data_class=FottLabel, data=fott_label_dict)
+            return fott_label